requirements.txt
# requirements.txt - Zero Rust dependencies
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
numpy==1.25.2
notion-client==2.2.1
python-dotenv==1.0.0


# That's it! No transformers, no tokenizers, no Rust compiler needed